fix: enforce float32 on load/query, deduplicate cache-clear into embedder
- Cast embeddings to float32 on load (astype with copy=False) to
  prevent float64 stores from corrupting the dtype invariant
- Use np.float32(1.0) in distance calculation to keep distances in
  float32 (Python 1.0 upcasts the result to float64)
- Make clear_model_cache() public in embedder.py and call it from
  init_repo.sh instead of duplicating the glob+rmtree logic inline

+++ src/engine/vector_store.py
@@ -61,7 +61,7 @@ class NumpyVectorStore:
         if os.path.exists(self._npz_path) and os.path.exists(self._meta_path):
             try:
                 with np.load(self._npz_path) as data:
-                    self._embeddings = data["embeddings"]
+                    self._embeddings = data["embeddings"].astype(np.float32, copy=False)
                     npz_version = str(data["save_version"]) if "save_version" in data.files else ""
 
                 with open(self._meta_path, "r", encoding="utf-8") as f:
@@ -343,7 +343,7 @@ class NumpyVectorStore:
             if self._normed is None:
                 self._recompute_norms()
             similarities = self._normed @ query_norm
-            distances = 1.0 - similarities
+            distances = np.float32(1.0) - similarities
 
             # Get top-n indices (use argsort for correctness at small N)
             n = min(n_results, len(self._ids))
